fix(omm): compile OMM modules against coopdx, exclude unported ones
- Add omm_geo_shim/surface_load_shim with coopdx-equivalent math, GBI,
  level script macros, level_cmd_get/level_script_preprocess, fs_* helpers
  and link stubs for excluded OMM UI/render/dialog/palette modules
- Fix omm_options, omm_rewards, omm_camera, omm_speedrun, omm_level,
  peachy/ and all object/model files to compile under coopdx
- Rewrite omm_stars as a save_file delegate; exclude omm_render_*/level
  custom levels/dialog_box/fs_packtype_zip from Android build

+++ data/omm/system/omm_audio.c
@@ -49,7 +49,7 @@ s32 omm_audio_resample_fast(u8 **output, const u8 *input, s32 inputLength, f32 o
 
     // Resampling
     for (s32 i = 0; i != outputSamples; ++i) {
-        temp[i] = inputBuffer[clamp_s(i * invOutputScale, 0, inputSamples - 1)];
+        temp[i] = inputBuffer[(s32) clamp_s(i * invOutputScale, 0, inputSamples - 1)];
     }
 
     // Copy temp to output
